[COST-863] - tech-debt: more obsolete code removal
