+++ koku/masu/database/aws_report_db_accessor.py
@@ -33,11 +33,7 @@ from reporting.models import OCPAWSCostLineItemProjectDailySummaryP
 from reporting.provider.aws.models import AWSCostEntry
 from reporting.provider.aws.models import AWSCostEntryBill
 from reporting.provider.aws.models import AWSCostEntryLineItem
-from reporting.provider.aws.models import AWSCostEntryLineItemDaily
 from reporting.provider.aws.models import AWSCostEntryLineItemDailySummary
-from reporting.provider.aws.models import AWSCostEntryPricing
-from reporting.provider.aws.models import AWSCostEntryProduct
-from reporting.provider.aws.models import AWSCostEntryReservation
 from reporting.provider.aws.models import PRESTO_LINE_ITEM_DAILY_TABLE
 from reporting.provider.aws.models import UI_SUMMARY_TABLES
 from reporting.provider.aws.openshift.models import UI_SUMMARY_TABLES as OCPAWS_UI_SUMMARY_TABLES
@@ -81,29 +77,6 @@ class AWSReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
     def cost_entry_table(self):
         return AWSCostEntry
 
-    @property
-    def line_item_daily_table(self):
-        return AWSCostEntryLineItemDaily
-
-    def get_cost_entry_bills(self):
-        """Get all cost entry bill objects."""
-        table_name = AWSCostEntryBill
-        with schema_context(self.schema):
-            columns = ["id", "bill_type", "payer_account_id", "billing_period_start", "provider_id"]
-            bills = self._get_db_obj_query(table_name).values(*columns)
-            return {
-                (bill["bill_type"], bill["payer_account_id"], bill["billing_period_start"], bill["provider_id"]): bill[
-                    "id"
-                ]
-                for bill in bills
-            }
-
-    def get_cost_entry_bills_by_date(self, start_date):
-        """Return a cost entry bill for the specified start date."""
-        table_name = AWSCostEntryBill
-        with schema_context(self.schema):
-            return self._get_db_obj_query(table_name).filter(billing_period_start=start_date)
-
     def get_cost_entry_bills_query_by_provider(self, provider_uuid):
         """Return all cost entry bills for the specified provider."""
         table_name = AWSCostEntryBill
@@ -130,140 +103,6 @@ class AWSReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
             else:
                 cost_entry_bill_query = base_query.filter(billing_period_start__lte=date)
             return cost_entry_bill_query
-
-    def get_lineitem_query_for_billid(self, bill_id):
-        """Get the AWS cost entry line item for a given bill query."""
-        table_name = AWSCostEntryLineItem
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-            return line_item_query
-
-    def get_daily_query_for_billid(self, bill_id):
-        """Get the AWS cost daily item for a given bill query."""
-        table_name = AWSCostEntryLineItemDaily
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            daily_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-            return daily_item_query
-
-    def get_summary_query_for_billid(self, bill_id):
-        """Get the AWS cost summary item for a given bill query."""
-        table_name = AWSCostEntryLineItemDailySummary
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            summary_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-            return summary_item_query
-
-    def get_ocp_aws_summary_query_for_billid(self, bill_id):
-        """Get the OCP-on-AWS report summary item for a given bill query."""
-        table_name = self._table_map["ocp_on_aws_daily_summary"]
-        base_query = self._get_db_obj_query(table_name)
-        summary_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-        return summary_item_query
-
-    def get_ocp_aws_project_summary_query_for_billid(self, bill_id):
-        """Get the OCP-on-AWS report project summary item for a given bill query."""
-        table_name = self._table_map["ocp_on_aws_project_daily_summary"]
-        base_query = self._get_db_obj_query(table_name)
-        summary_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-        return summary_item_query
-
-    def get_cost_entry_query_for_billid(self, bill_id):
-        """Get the AWS cost entry data for a given bill query."""
-        table_name = AWSCostEntry
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(bill_id=bill_id)
-            return line_item_query
-
-    def get_cost_entries(self):
-        """Make a mapping of cost entries by start time."""
-        table_name = AWSCostEntry
-        with schema_context(self.schema):
-            cost_entries = self._get_db_obj_query(table_name).all()
-
-            return {(ce.bill_id, ce.interval_start.strftime(self._datetime_format)): ce.id for ce in cost_entries}
-
-    def get_products(self):
-        """Make a mapping of product sku to product objects."""
-        table_name = AWSCostEntryProduct
-        with schema_context(self.schema):
-            columns = ["id", "sku", "product_name", "region"]
-            products = self._get_db_obj_query(table_name, columns=columns).all()
-
-            return {
-                (product["sku"], product["product_name"], product["region"]): product["id"] for product in products
-            }
-
-    def get_pricing(self):
-        """Make a mapping of pricing values string to pricing objects."""
-        table_name = AWSCostEntryPricing
-        with schema_context(self.schema):
-            pricing = self._get_db_obj_query(table_name).all()
-
-            return {f"{p.term}-{p.unit}": p.id for p in pricing}
-
-    def get_reservations(self):
-        """Make a mapping of reservation ARN to reservation objects."""
-        table_name = AWSCostEntryReservation
-        with schema_context(self.schema):
-            columns = ["id", "reservation_arn"]
-            reservs = self._get_db_obj_query(table_name, columns=columns).all()
-
-            return {res["reservation_arn"]: res["id"] for res in reservs}
-
-    def populate_line_item_daily_table(self, start_date, end_date, bill_ids):
-        """Populate the daily aggregate of line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-            bill_ids (list)
-
-        Returns
-            (None)
-
-        """
-        table_name = self._table_map["line_item_daily"]
-
-        daily_sql = pkgutil.get_data("masu.database", "sql/reporting_awscostentrylineitem_daily.sql")
-        daily_sql = daily_sql.decode("utf-8")
-        daily_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "bill_ids": bill_ids,
-            "schema": self.schema,
-        }
-        daily_sql, daily_sql_params = self.jinja_sql.prepare_query(daily_sql, daily_sql_params)
-        self._execute_raw_sql_query(table_name, daily_sql, start_date, end_date, bind_params=list(daily_sql_params))
-
-    def populate_line_item_daily_summary_table(self, start_date, end_date, bill_ids):
-        """Populate the daily aggregated summary of line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-
-        Returns
-            (None)
-
-        """
-        table_name = self._table_map["line_item_daily_summary"]
-        summary_sql = pkgutil.get_data("masu.database", "sql/reporting_awscostentrylineitem_daily_summary.sql")
-        summary_sql = summary_sql.decode("utf-8")
-        summary_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "bill_ids": bill_ids,
-            "schema": self.schema,
-        }
-        summary_sql, summary_sql_params = self.jinja_sql.prepare_query(summary_sql, summary_sql_params)
-        self._execute_raw_sql_query(
-            table_name, summary_sql, start_date, end_date, bind_params=list(summary_sql_params)
-        )
 
     def populate_ui_summary_tables(self, start_date, end_date, source_uuid, tables=UI_SUMMARY_TABLES):
         """Populate our UI summary tables (formerly materialized views)."""
@@ -337,35 +176,6 @@ class AWSReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
         agg_sql_params = {"schema": self.schema, "bill_ids": bill_ids, "start_date": start_date, "end_date": end_date}
         agg_sql, agg_sql_params = self.jinja_sql.prepare_query(agg_sql, agg_sql_params)
         self._execute_raw_sql_query(table_name, agg_sql, bind_params=list(agg_sql_params))
-
-    def populate_ocp_on_aws_cost_daily_summary(self, start_date, end_date, cluster_id, bill_ids, markup_value):
-        """Populate the daily cost aggregated summary for OCP on AWS.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-
-        Returns
-            (None)
-
-        """
-        table_name = self._table_map["ocp_on_aws_daily_summary"]
-        summary_sql = pkgutil.get_data("masu.database", "sql/reporting_ocpawscostlineitem_daily_summary.sql")
-        summary_sql = summary_sql.decode("utf-8")
-        summary_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "bill_ids": bill_ids,
-            "cluster_id": cluster_id,
-            "schema": self.schema,
-            "markup": markup_value,
-        }
-        summary_sql, summary_sql_params = self.jinja_sql.prepare_query(summary_sql, summary_sql_params)
-
-        self._execute_raw_sql_query(
-            table_name, summary_sql, start_date, end_date, bind_params=list(summary_sql_params)
-        )
 
     def populate_ocp_on_aws_ui_summary_tables(self, sql_params, tables=OCPAWS_UI_SUMMARY_TABLES):
         """Populate our UI summary tables (formerly materialized views)."""
