+++ koku/masu/database/azure_report_db_accessor.py
@@ -7,7 +7,6 @@ import json
 import logging
 import pkgutil
 import uuid
-from datetime import datetime
 
 from dateutil.parser import parse
 from django.conf import settings
@@ -31,10 +30,7 @@ from reporting.models import OCPAllCostLineItemDailySummaryP
 from reporting.models import OCPAllCostLineItemProjectDailySummaryP
 from reporting.models import OCPAzureCostLineItemProjectDailySummaryP
 from reporting.provider.azure.models import AzureCostEntryBill
-from reporting.provider.azure.models import AzureCostEntryLineItemDaily
 from reporting.provider.azure.models import AzureCostEntryLineItemDailySummary
-from reporting.provider.azure.models import AzureCostEntryProductService
-from reporting.provider.azure.models import AzureMeter
 from reporting.provider.azure.models import PRESTO_LINE_ITEM_TABLE
 from reporting.provider.azure.models import UI_SUMMARY_TABLES
 from reporting.provider.azure.openshift.models import UI_SUMMARY_TABLES as OCPAZURE_UI_SUMMARY_TABLES
@@ -70,44 +66,6 @@ class AzureReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
     def ocpall_line_item_project_daily_summary_table(self):
         return get_model("OCPAllCostLineItemProjectDailySummaryP")
 
-    @property
-    def line_item_daily_table(self):
-        return AzureCostEntryLineItemDaily
-
-    def get_cost_entry_bills(self):
-        """Get all cost entry bill objects."""
-        table_name = AzureCostEntryBill
-        with schema_context(self.schema):
-            columns = ["id", "billing_period_start", "provider_id"]
-            bills = self._get_db_obj_query(table_name).values(*columns)
-            return {(bill["billing_period_start"], bill["provider_id"]): bill["id"] for bill in bills}
-
-    def get_products(self):
-        """Make a mapping of product objects."""
-        table_name = AzureCostEntryProductService
-        with schema_context(self.schema):
-            columns = ["id", "instance_id", "instance_type", "service_name", "service_tier"]
-            products = self._get_db_obj_query(table_name, columns=columns).all()
-
-            return {
-                (
-                    product["instance_id"],
-                    product["instance_type"],
-                    product["service_tier"],
-                    product["service_name"],
-                ): product["id"]
-                for product in products
-            }
-
-    def get_meters(self):
-        """Make a mapping of meter objects."""
-        table_name = AzureMeter
-        with schema_context(self.schema):
-            columns = ["id", "meter_id"]
-            meters = self._get_db_obj_query(table_name, columns=columns).all()
-
-            return {(meter["meter_id"]): meter["id"] for meter in meters}
-
     def get_cost_entry_bills_query_by_provider(self, provider_uuid):
         """Return all cost entry bills for the specified provider."""
         table_name = AzureCostEntryBill
@@ -123,36 +81,6 @@ class AzureReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
             bill_date = start_date.replace(day=1)
             bills = bills.filter(billing_period_start=bill_date)
         return bills
-
-    def populate_line_item_daily_summary_table(self, start_date, end_date, bill_ids):
-        """Populate the daily aggregated summary of line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-
-        Returns
-            (None)
-
-        """
-
-        _start_date = start_date.date() if isinstance(start_date, datetime) else start_date
-        _end_date = end_date.date() if isinstance(end_date, datetime) else end_date
-
-        table_name = self._table_map["line_item_daily_summary"]
-        summary_sql = pkgutil.get_data("masu.database", "sql/reporting_azurecostentrylineitem_daily_summary.sql")
-        summary_sql = summary_sql.decode("utf-8")
-        summary_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": _start_date,
-            "end_date": _end_date,
-            "bill_ids": bill_ids,
-            "schema": self.schema,
-        }
-        summary_sql, summary_sql_params = self.jinja_sql.prepare_query(summary_sql, summary_sql_params)
-        self._execute_raw_sql_query(
-            table_name, summary_sql, start_date, end_date, bind_params=list(summary_sql_params)
-        )
 
     def populate_line_item_daily_summary_table_presto(self, start_date, end_date, source_uuid, bill_id, markup_value):
         """Populate the daily aggregated summary of line items table.
@@ -198,12 +126,6 @@ class AzureReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
         agg_sql, agg_sql_params = self.jinja_sql.prepare_query(agg_sql, agg_sql_params)
         self._execute_raw_sql_query(table_name, agg_sql, bind_params=list(agg_sql_params))
 
-    def get_cost_entry_bills_by_date(self, start_date):
-        """Return a cost entry bill for the specified start date."""
-        table_name = AzureCostEntryBill
-        with schema_context(self.schema):
-            return self._get_db_obj_query(table_name).filter(billing_period_start=start_date)
-
     def populate_markup_cost(self, provider_uuid, markup, start_date, end_date, bill_ids=None):
         """Set markup costs in the database."""
         with schema_context(self.schema):
@@ -245,22 +167,6 @@ class AzureReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
             else:
                 cost_entry_bill_query = base_query.filter(billing_period_start__lte=date)
             return cost_entry_bill_query
-
-    def get_lineitem_query_for_billid(self, bill_id):
-        """Get the Azure cost entry line item for a given bill query."""
-        table_name = AzureCostEntryLineItemDaily
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-            return line_item_query
-
-    def get_summary_query_for_billid(self, bill_id):
-        """Get the Azure cost summary item for a given bill query."""
-        table_name = AzureCostEntryLineItemDailySummary
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            summary_item_query = base_query.filter(cost_entry_bill_id=bill_id)
-            return summary_item_query
 
     def populate_ocp_on_azure_cost_daily_summary(self, start_date, end_date, cluster_id, bill_ids, markup_value):
         """Populate the daily cost aggregated summary for OCP on Azure.
