+++ koku/masu/database/ocp_report_db_accessor.py
@@ -50,8 +50,6 @@ from reporting.provider.ocp.models import OCPNode
 from reporting.provider.ocp.models import OCPProject
 from reporting.provider.ocp.models import OCPPVC
 from reporting.provider.ocp.models import OCPUsageLineItemDailySummary
-from reporting.provider.ocp.models import OCPUsageReport
-from reporting.provider.ocp.models import OCPUsageReportPeriod
 from reporting.provider.ocp.models import PRESTO_LINE_ITEM_TABLE_DAILY_MAP
 from reporting.provider.ocp.models import UI_SUMMARY_TABLES
 
@@ -93,25 +91,12 @@ class OCPReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
     def line_item_daily_summary_table(self):
         return OCPUsageLineItemDailySummary
 
-    def get_current_usage_report(self):
-        """Get the most recent usage report object."""
-        table_name = self._table_map["report"]
-
-        with schema_context(self.schema):
-            return self._get_db_obj_query(table_name).order_by("-interval_start").first()
-
     def get_current_usage_period(self):
         """Get the most recent usage report period object."""
         table_name = self._table_map["report_period"]
 
         with schema_context(self.schema):
             return self._get_db_obj_query(table_name).order_by("-report_period_start").first()
-
-    def get_usage_periods_by_date(self, start_date):
-        """Return all report period entries for the specified start date."""
-        table_name = self._table_map["report_period"]
-        with schema_context(self.schema):
-            return self._get_db_obj_query(table_name).filter(report_period_start=start_date).all()
 
     def get_usage_period_by_dates_and_cluster(self, start_date, end_date, cluster_id):
         """Return all report period entries for the specified start date."""
@@ -122,18 +107,6 @@ class OCPReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
                 .filter(report_period_start=start_date, report_period_end=end_date, cluster_id=cluster_id)
                 .first()
             )
-
-    def get_usage_period_on_or_before_date(self, date, provider_uuid=None):
-        """Get the usage report period objects before provided date."""
-        table_name = self._table_map["report_period"]
-
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            if provider_uuid:
-                usage_period_query = base_query.filter(report_period_start__lte=date, provider_id=provider_uuid)
-            else:
-                usage_period_query = base_query.filter(report_period_start__lte=date)
-            return usage_period_query
 
     def get_usage_period_query_by_provider(self, provider_uuid):
         """Return all report periods for the specified provider."""
@@ -152,212 +125,6 @@ class OCPReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
                 report_periods = report_periods.filter(report_period_start=report_date).first()
 
             return report_periods
-
-    def get_lineitem_query_for_reportid(self, query_report_id):
-        """Get the usage report line item for a report id query."""
-        table_name = self._table_map["line_item"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(report_id=query_report_id)
-            return line_item_query
-
-    def get_daily_usage_query_for_clusterid(self, cluster_identifier):
-        """Get the usage report daily item for a cluster id query."""
-        table_name = self._table_map["line_item_daily"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            daily_usage_query = base_query.filter(cluster_id=cluster_identifier)
-            return daily_usage_query
-
-    def get_summary_usage_query_for_clusterid(self, cluster_identifier):
-        """Get the usage report summary for a cluster id query."""
-        table_name = self._table_map["line_item_daily_summary"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            summary_usage_query = base_query.filter(cluster_id=cluster_identifier)
-            return summary_usage_query
-
-    def get_item_query_report_period_id(self, report_period_id):
-        """Get the usage report line item for a report id query."""
-        table_name = self._table_map["line_item"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(report_period_id=report_period_id)
-            return line_item_query
-
-    def get_storage_item_query_report_period_id(self, report_period_id):
-        """Get the storage report line item for a report id query."""
-        table_name = self._table_map["storage_line_item"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(report_period_id=report_period_id)
-            return line_item_query
-
-    def get_daily_storage_item_query_cluster_id(self, cluster_identifier):
-        """Get the daily storage report line item for a cluster id query."""
-        table_name = self._table_map["storage_line_item_daily"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            daily_item_query = base_query.filter(cluster_id=cluster_identifier)
-            return daily_item_query
-
-    def get_storage_summary_query_cluster_id(self, cluster_identifier):
-        """Get the storage report summary for a cluster id query."""
-        table_name = self._table_map["line_item_daily_summary"]
-        filters = {"cluster_id": cluster_identifier, "data_source": "Storage"}
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            daily_item_query = base_query.filter(**filters)
-            return daily_item_query
-
-    def get_node_label_item_query_report_period_id(self, report_period_id):
-        """Get the node label report line item for a report id query."""
-        table_name = self._table_map["node_label_line_item"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            line_item_query = base_query.filter(report_period_id=report_period_id)
-            return line_item_query
-
-    def get_ocp_aws_summary_query_for_cluster_id(self, cluster_identifier):
-        """Get the OCP-on-AWS report summary item for a given cluster id query."""
-        table_name = self._aws_table_map["ocp_on_aws_daily_summary"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            summary_item_query = base_query.filter(cluster_id=cluster_identifier)
-            return summary_item_query
-
-    def get_ocp_aws_project_summary_query_for_cluster_id(self, cluster_identifier):
-        """Get the OCP-on-AWS report project summary item for a given cluster id query."""
-        table_name = self._aws_table_map["ocp_on_aws_project_daily_summary"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            summary_item_query = base_query.filter(cluster_id=cluster_identifier)
-            return summary_item_query
-
-    def get_report_query_report_period_id(self, report_period_id):
-        """Get the usage report line item for a report id query."""
-        table_name = self._table_map["report"]
-        with schema_context(self.schema):
-            base_query = self._get_db_obj_query(table_name)
-            usage_report_query = base_query.filter(report_period_id=report_period_id)
-            return usage_report_query
-
-    def get_report_periods(self):
-        """Get all usage period objects."""
-        periods = []
-        with schema_context(self.schema):
-            periods = OCPUsageReportPeriod.objects.values("id", "cluster_id", "report_period_start", "provider_id")
-            return_value = {(p["cluster_id"], p["report_period_start"], p["provider_id"]): p["id"] for p in periods}
-            return return_value
-
-    def get_reports(self):
-        """Make a mapping of reports by time."""
-        with schema_context(self.schema):
-            reports = OCPUsageReport.objects.all()
-            return {
-                (entry.report_period_id, entry.interval_start.strftime(self._datetime_format)): entry.id
-                for entry in reports
-            }
-
-    def get_pod_usage_cpu_core_hours(self, start_date, end_date, cluster_id=None):
-        """Make a mapping of cpu pod usage hours."""
-        table = OCPUsageLineItemDailySummary
-        filters = create_filter("Pod", start_date, end_date, cluster_id)
-        with schema_context(self.schema):
-            reports = self._get_reports(table, filters)
-            return {entry.uuid: entry.pod_usage_cpu_core_hours for entry in reports}
-
-    def _get_reports(self, table, filters=None):
-        """Return requested reports from given table.
-
-        Args:
-            table (Django models.Model object): The table to query against
-            filters (dict): Columns to filter the query on
-
-        Returns:
-            (QuerySet): Django queryset of objects queried on
-
-        """
-        with schema_context(self.schema):
-            if filters:
-                reports = self._get_db_obj_query(table).filter(**filters).all()
-            else:
-                reports = self._get_db_obj_query(table).all()
-            return reports
-
-    def get_pod_request_cpu_core_hours(self, start_date, end_date, cluster_id=None):
-        """Make a mapping of cpu pod request hours."""
-        table = OCPUsageLineItemDailySummary
-        filters = create_filter("Pod", start_date, end_date, cluster_id)
-        with schema_context(self.schema):
-            reports = self._get_reports(table, filters)
-            return {entry.uuid: entry.pod_request_cpu_core_hours for entry in reports}
-
-    def get_pod_usage_memory_gigabyte_hours(self, start_date, end_date, cluster_id=None):
-        """Make a mapping of memory_usage hours."""
-        table = OCPUsageLineItemDailySummary
-        filters = create_filter("Pod", start_date, end_date, cluster_id)
-        with schema_context(self.schema):
-            reports = self._get_reports(table, filters)
-            return {entry.uuid: entry.pod_usage_memory_gigabyte_hours for entry in reports}
-
-    def get_pod_request_memory_gigabyte_hours(self, start_date, end_date, cluster_id=None):
-        """Make a mapping of memory_request_hours."""
-        table = OCPUsageLineItemDailySummary
-        filters = create_filter("Pod", start_date, end_date, cluster_id)
-        with schema_context(self.schema):
-            reports = self._get_reports(table, filters)
-            return {entry.uuid: entry.pod_request_memory_gigabyte_hours for entry in reports}
-
-    def get_persistentvolumeclaim_usage_gigabyte_months(self, start_date, end_date, cluster_id=None):
-        """Make a mapping of persistentvolumeclaim_usage_gigabyte_months."""
-        table = OCPUsageLineItemDailySummary
-        filters = create_filter("Storage", start_date, end_date, cluster_id)
-        with schema_context(self.schema):
-            reports = self._get_reports(table, filters)
-            return {entry.uuid: entry.persistentvolumeclaim_usage_gigabyte_months for entry in reports}
-
-    def get_volume_request_storage_gigabyte_months(self, start_date, end_date, cluster_id=None):
-        """Make a mapping of volume_request_storage_gigabyte_months."""
-        table = OCPUsageLineItemDailySummary
-        filters = create_filter("Storage", start_date, end_date, cluster_id)
-        with schema_context(self.schema):
-            reports = self._get_reports(table, filters)
-            return {entry.uuid: entry.volume_request_storage_gigabyte_months for entry in reports}
-
-    def populate_line_item_daily_table(self, start_date, end_date, cluster_id):
-        """Populate the daily aggregate of line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-            cluster_id (String) Cluster Identifier
-
-        Returns
-            (None)
-
-        """
-        # Cast start_date and end_date into date object instead of string
-        if isinstance(start_date, str):
-            start_date = datetime.datetime.strptime(start_date, "%Y-%m-%d").date()
-            end_date = datetime.datetime.strptime(end_date, "%Y-%m-%d").date()
-        if isinstance(start_date, datetime.datetime):
-            start_date = start_date.date()
-            end_date = end_date.date()
-
-        table_name = self._table_map["line_item_daily"]
-
-        daily_sql = pkgutil.get_data("masu.database", "sql/reporting_ocpusagelineitem_daily.sql")
-        daily_sql = daily_sql.decode("utf-8")
-        daily_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "cluster_id": cluster_id,
-            "schema": self.schema,
-        }
-        daily_sql, daily_sql_params = self.jinja_sql.prepare_query(daily_sql, daily_sql_params)
-        self._execute_raw_sql_query(table_name, daily_sql, start_date, end_date, bind_params=list(daily_sql_params))
 
     def populate_ui_summary_tables(self, start_date, end_date, source_uuid, tables=UI_SUMMARY_TABLES):
         """Populate our UI summary tables (formerly materialized views)."""
@@ -530,148 +297,6 @@ class OCPReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
             db_results[entry[0]] = (entry[1], entry[2])
         return db_results
 
-    def populate_storage_line_item_daily_table(self, start_date, end_date, cluster_id):
-        """Populate the daily storage aggregate of line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-            cluster_id (String) Cluster Identifier
-
-        Returns
-            (None)
-
-        """
-        # Cast string to date object
-        if isinstance(start_date, str):
-            start_date = datetime.datetime.strptime(start_date, "%Y-%m-%d").date()
-            end_date = datetime.datetime.strptime(end_date, "%Y-%m-%d").date()
-        if isinstance(start_date, datetime.datetime):
-            start_date = start_date.date()
-            end_date = end_date.date()
-        table_name = self._table_map["storage_line_item_daily"]
-
-        daily_sql = pkgutil.get_data("masu.database", "sql/reporting_ocpstoragelineitem_daily.sql")
-        daily_sql = daily_sql.decode("utf-8")
-        daily_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "cluster_id": cluster_id,
-            "schema": self.schema,
-        }
-        daily_sql, daily_sql_params = self.jinja_sql.prepare_query(daily_sql, daily_sql_params)
-        self._execute_raw_sql_query(table_name, daily_sql, start_date, end_date, bind_params=list(daily_sql_params))
-
-    def populate_pod_charge(self, cpu_temp_table, mem_temp_table):
-        """Populate the memory and cpu charge on daily summary table.
-
-        Args:
-            cpu_temp_table (String) Name of cpu charge temp table
-            mem_temp_table (String) Name of mem charge temp table
-
-        Returns
-            (None)
-
-        """
-        table_name = self._table_map["line_item_daily_summary"]
-
-        daily_charge_sql = pkgutil.get_data("masu.database", "sql/reporting_ocpusagelineitem_daily_pod_charge.sql")
-        charge_line_sql = daily_charge_sql.decode("utf-8")
-        charge_line_sql_params = {"cpu_temp": cpu_temp_table, "mem_temp": mem_temp_table, "schema": self.schema}
-        charge_line_sql, charge_line_sql_params = self.jinja_sql.prepare_query(charge_line_sql, charge_line_sql_params)
-        self._execute_raw_sql_query(table_name, charge_line_sql, bind_params=list(charge_line_sql_params))
-
-    def populate_storage_charge(self, temp_table_name):
-        """Populate the storage charge into the daily summary table.
-
-        Args:
-            storage_charge (Float) Storage charge.
-
-        Returns
-            (None)
-
-        """
-        table_name = self._table_map["line_item_daily_summary"]
-
-        daily_charge_sql = pkgutil.get_data("masu.database", "sql/reporting_ocp_storage_charge.sql")
-        charge_line_sql = daily_charge_sql.decode("utf-8")
-        charge_line_sql_params = {"temp_table": temp_table_name, "schema": self.schema}
-        charge_line_sql, charge_line_sql_params = self.jinja_sql.prepare_query(charge_line_sql, charge_line_sql_params)
-        self._execute_raw_sql_query(table_name, charge_line_sql, bind_params=list(charge_line_sql_params))
-
-    def populate_line_item_daily_summary_table(self, start_date, end_date, cluster_id, source):
-        """Populate the daily aggregate of line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-            cluster_id (String) Cluster Identifier
-            source (String) Source UUID
-
-        Returns
-            (None)
-
-        """
-        # Cast start_date to date
-        if isinstance(start_date, str):
-            start_date = datetime.datetime.strptime(start_date, "%Y-%m-%d").date()
-            end_date = datetime.datetime.strptime(end_date, "%Y-%m-%d").date()
-        if isinstance(start_date, datetime.datetime):
-            start_date = start_date.date()
-            end_date = end_date.date()
-        table_name = self._table_map["line_item_daily_summary"]
-
-        summary_sql = pkgutil.get_data("masu.database", "sql/reporting_ocpusagelineitem_daily_summary.sql")
-        summary_sql = summary_sql.decode("utf-8")
-        summary_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "cluster_id": cluster_id,
-            "schema": self.schema,
-            "source_uuid": source,
-        }
-        summary_sql, summary_sql_params = self.jinja_sql.prepare_query(summary_sql, summary_sql_params)
-        self._execute_raw_sql_query(
-            table_name, summary_sql, start_date, end_date, bind_params=list(summary_sql_params)
-        )
-
-    def populate_storage_line_item_daily_summary_table(self, start_date, end_date, cluster_id, source):
-        """Populate the daily aggregate of storage line items table.
-
-        Args:
-            start_date (datetime.date) The date to start populating the table.
-            end_date (datetime.date) The date to end on.
-            cluster_id (String) Cluster Identifier
-            source (String) Source UUID
-
-        Returns
-            (None)
-
-        """
-        # Cast start_date and end_date to date object, if they aren't already
-        if isinstance(start_date, str):
-            start_date = datetime.datetime.strptime(start_date, "%Y-%m-%d").date()
-            end_date = datetime.datetime.strptime(end_date, "%Y-%m-%d").date()
-        if isinstance(start_date, datetime.datetime):
-            start_date = start_date.date()
-            end_date = end_date.date()
-        table_name = self._table_map["line_item_daily_summary"]
-
-        summary_sql = pkgutil.get_data("masu.database", "sql/reporting_ocpstoragelineitem_daily_summary.sql")
-        summary_sql = summary_sql.decode("utf-8")
-        summary_sql_params = {
-            "uuid": str(uuid.uuid4()).replace("-", "_"),
-            "start_date": start_date,
-            "end_date": end_date,
-            "cluster_id": cluster_id,
-            "schema": self.schema,
-            "source_uuid": source,
-        }
-        summary_sql, summary_sql_params = self.jinja_sql.prepare_query(summary_sql, summary_sql_params)
-        self._execute_raw_sql_query(table_name, summary_sql, start_date, end_date, list(summary_sql_params))
-
     def delete_ocp_hive_partition_by_day(self, days, source, year, month):
         """Deletes partitions individually for each day in days list."""
         table = "reporting_ocpusagelineitem_daily_summary"
@@ -824,13 +449,6 @@ class OCPReportDBAccessor(SQLScriptAtomicExecutorMixin, ReportDBAccessorBase):
         finally:
             LOG.info("TRINO OCP: Close connection")
             presto_conn.close()
-
-    def get_cost_summary_for_clusterid(self, cluster_identifier):
-        """Get the cost summary for a cluster id query."""
-        table_name = self._table_map["cost_summary"]
-        base_query = self._get_db_obj_query(table_name)
-        cost_summary_query = base_query.filter(cluster_id=cluster_identifier)
-        return cost_summary_query
 
     def populate_pod_label_summary_table(self, report_period_ids, start_date, end_date):
         """Populate the line item aggregated totals data table."""
