+++ koku/masu/processor/_tasks/process.py
@@ -4,7 +4,6 @@
 #
 """Asynchronous tasks."""
 import logging
-from os import path
 
 import psutil
 
@@ -88,9 +87,6 @@ def _process_report_file(schema_name, provider, report_dict):
             LOG.error(log_json(tracing_id, ("Unable to find manifest for ID: %s, file %s", manifest_id, file_name)))
 
     with ProviderDBAccessor(provider_uuid=provider_uuid) as provider_accessor:
-        if provider_accessor.get_setup_complete():
-            files = processor.remove_processed_files(path.dirname(report_path))
-            LOG.info(log_json(tracing_id, ("Temporary files removed: %s", str(files))))
         provider_accessor.setup_complete()
 
     return True
