+++ koku/masu/processor/ocp/ocp_report_db_cleaner.py
@@ -92,7 +92,6 @@ class OCPReportDBCleaner:
         all_period_starts = set()
 
         with OCPReportDBAccessor(self._schema) as accessor:
-            # all_usage_periods = accessor.get_usage_periods_by_date(expired_date)
             all_usage_periods = accessor._get_db_obj_query(accessor._table_map["report_period"]).filter(
                 report_period_start__lte=expired_date
             )
